# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'EventHub Core API'
  description: 'REST API for the EventHub multi-vendor event ticketing and payout platform. Vendors create events and sell tickets; attendees browse and buy; admins approve vendors, manage refunds, and disburse payouts. Every financial operation is auditable, idempotent, and resilient to partial failure.'
  version: 1.0.0
servers:
  - url: 'http://localhost'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'Obtain a token via **POST /api/v1/auth/login**. Include it in every authenticated request: `Authorization: Bearer {token}`.'
security:
  -
    default: []
paths:
...
  /api/v1/orders:
    post:
      summary: Checkout
      operationId: checkout
      description: "Reserve tickets with a 15-minute hold and initiate payment via the payment-service.\nThe order is created with `status=pending`; it becomes `paid` when the payment webhook\nconfirms success, or `expired` if the hold expires without payment.\n\n**Idempotency:** include a unique `Idempotency-Key` header. Replaying the same key returns\nthe existing order without creating a second charge."
      parameters:
        -
          in: header
          name: Idempotency-Key
          description: ''
          example: 'string required A unique key (UUID recommended) to make this request idempotent. Example: 550e8400-e29b-41d4-a716-446655440000'
          schema:
            type: string
      responses:
        201:
          description: 'Order created (pending payment)'
          content:
            application/json:
              schema:
                type: object
                example:
                  success: true
                  message: 'Order created, payment initiated.'
                  data:
                    order:
                      id: 01J000000000000DEMOORDER1
                      status:
                        value: pending
                        label: Pending
                      total: 75000
                      currency: BDT
                      items:
                        -
                          ticket_type_id: 01J000000000000DEMOTICKET
                          quantity: 3
                          unit_price: 25000
                      created_at: '2026-06-30T10:05:00Z'
                  errors: null
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Order created, payment initiated.'
                  data:
                    type: object
                    properties:
                      order:
                        type: object
                        properties:
                          id:
                            type: string
                            example: 01J000000000000DEMOORDER1
                          status:
                            type: object
                            properties:
                              value:
                                type: string
                                example: pending
                              label:
                                type: string
                                example: Pending
                          total:
                            type: integer
                            example: 75000
                          currency:
                            type: string
                            example: BDT
                          items:
                            type: array
                            example:
                              - ticket_type_id: 01J000000000000DEMOTICKET
                                quantity: 3
                                unit_price: 25000
                            items:
                              type: object
                              properties:
                                ticket_type_id:
                                  type: string
                                  example: 01J000000000000DEMOTICKET
                                quantity:
                                  type: integer
                                  example: 3
                                unit_price:
                                  type: integer
                                  example: 25000
                          created_at:
                            type: string
                            example: '2026-06-30T10:05:00Z'
                  errors:
                    type: string
                    example: null
                    nullable: true
        409:
          description: 'Tickets unavailable'
          content:
            application/json:
              schema:
                type: object
                example:
                  success: false
                  message: 'Insufficient tickets available. Please try a smaller quantity or a different ticket type.'
                  data: null
                  errors: null
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Insufficient tickets available. Please try a smaller quantity or a different ticket type.'
                  data:
                    type: string
                    example: null
                    nullable: true
                  errors:
                    type: string
                    example: null
                    nullable: true
        422:
          description: 'Missing idempotency key'
          content:
            application/json:
              schema:
                type: object
                example:
                  success: false
                  message: 'Validation failed.'
                  data: null
                  errors:
                    idempotency_key:
                      - 'The idempotency key is required.'
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Validation failed.'
                  data:
                    type: string
                    example: null
                    nullable: true
                  errors:
                    type: object
                    properties:
                      idempotency_key:
                        type: array
                        example:
                          - 'The idempotency key is required.'
                        items:
                          type: string
      tags:
        - 'Orders / Checkout'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                idempotency_key:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: b
                items:
                  type: array
                  description: 'Must have at least 1 items. Must not have more than 50 items.'
                  example:
                    - []
                  items:
                    type: object
                    properties:
                      ticket_type_id:
                        type: string
                        description: 'Must match an existing stored value.'
                        example: architecto
                      quantity:
                        type: integer
                        description: 'Must be at least 1. Must not be greater than 100.'
                        example: 22
                    required:
                      - ticket_type_id
                      - quantity
              required:
                - idempotency_key
                - items
...
  '/api/v1/admin/payouts/{payout_id}/execute':
    post:
      summary: 'Execute payout'
      operationId: executePayout
      description: "Dispatch payment for a `pending` or `approved` payout. The job transitions the payout to\n`processing` and calls payment-service; the terminal result (`paid` or `failed`) arrives\nvia the signed payment-service webhook. Idempotent — re-dispatching the same payout\nreuses the deterministic idempotency key."
      parameters: []
      responses:
        200:
          description: Queued
          content:
            application/json:
              schema:
                type: object
                example:
                  success: true
                  message: 'Payout execution queued.'
                  data:
                    payout:
                      id: 01J000000000000DEMOPAYOUT
                      vendor_id: 01J000000000000DEMOVENDOR
                      gross: 500000
                      commission: 50000
                      net: 450000
                      currency: BDT
                      status:
                        value: pending
                        label: Pending
                      batch_id: 01J0BATCHID
                      created_at: '2026-06-30T09:00:00Z'
                  errors: null
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Payout execution queued.'
                  data:
                    type: object
                    properties:
                      payout:
                        type: object
                        properties:
                          id:
                            type: string
                            example: 01J000000000000DEMOPAYOUT
                          vendor_id:
                            type: string
                            example: 01J000000000000DEMOVENDOR
                          gross:
                            type: integer
                            example: 500000
                          commission:
                            type: integer
                            example: 50000
                          net:
                            type: integer
                            example: 450000
                          currency:
                            type: string
                            example: BDT
                          status:
                            type: object
                            properties:
                              value:
                                type: string
                                example: pending
                              label:
                                type: string
                                example: Pending
                          batch_id:
                            type: string
                            example: 01J0BATCHID
                          created_at:
                            type: string
                            example: '2026-06-30T09:00:00Z'
                  errors:
                    type: string
                    example: null
                    nullable: true
        409:
          description: 'Not executable'
          content:
            application/json:
              schema:
                type: object
                example:
                  success: false
                  message: 'This payout cannot be executed in its current status.'
                  data: null
                  errors: null
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'This payout cannot be executed in its current status.'
                  data:
                    type: string
                    example: null
                    nullable: true
                  errors:
                    type: string
                    example: null
                    nullable: true
      tags:
        - Payouts
    parameters:
      -
        in: path
        name: payout_id
        description: 'The ID of the payout.'
        example: architecto
        required: true
        schema:
          type: string